requirements.txt
dtw-python==1.1.10
tqdm==4.60.0
numpy==1.21.4
pandas==1.3.4
matplotlib==3.4.1
editdistance
